requirements.txt
requests
python-dotenv
schema
